pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-store-mcp"
version = "0.6.0"
description = "Open-source MCP tools for AI agents — lint, security, dependency audit, web scraping and more"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["mcp", "ai", "agents", "tools", "lint", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp>=1.0",
    "keyring>=25.0",
    "PyGithub>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
]

[project.scripts]
agentic-store-mcp = "agentic_store_mcp.server:main"
agentic-store-webapp = "agentic_store_mcp.webapp.run:main"

[project.urls]
Homepage = "https://www.agenticstore.dev/docs"
Issues   = "https://github.com/agenticstore/agentic-store-mcp/issues"
Repository = "https://github.com/agenticstore/agentic-store-mcp"


[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build]
exclude = [
  "*.pyc",
  "**/__pycache__",
  "**/.venv",
  "**/venv",
  "**/testing",
  "**/.pytest_cache",
  "**/.mypy_cache",
  "**/.ruff_cache",
  "**/node_modules",
  "dist",
  "build",
  "memory",
  "skills",
  ".env",
]

[tool.hatch.build.targets.wheel]
packages = ["agentic_store_mcp"]

[tool.hatch.build.targets.sdist]
include = [
  "agentic_store_mcp/",
  "server.py",
  "webapp.py",
  "setup_local.py",
  "README.md",
  "pyproject.toml",
  ".env.example",
]

[tool.pytest.ini_options]
testpaths = ["agentic_store_mcp/modules"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--import-mode=importlib"

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
